day 16 part 1
the struggle is real. I invested 3 hours trying to get it done recursively. It happily returned the correct result with the sample, but not for the input.

i converted the credited version to typescript and this one works like a charm, but i don't see my thinking problem, why it doesn't work in the recursive version...